feat: client-side controls testing
This script will be populated with user-specific information and then
distributred via the dashboard (which will of course require
authentication of the user).

The advantage of this script is that the user's credentials will never
have to leave their servers. In other words, by using this, we can help
our users without having to access, transmit, or store any of the user's
sensitive information, such as GRB data (customer information) and PCI
data (card numbers), or any of the credentials to other third parties,
which in this case is Supabase.

Signed-off-by: Soobin Rho <soobinrho@gmail.com>